+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [UNRELEASED]
 
 - added ability to pin tasks to the top of the task sidebar
+- added AI-powered task name generation
 
 ## [0.46.0]